project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "churn-mlops"
version = "0.1.0"
description = "Customer Churn Prediction - MLOps Course Project"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MLOps Course", email = "mlops@example.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "numpy>=1.21.0,<2.0.0",
    "pandas>=1.3.0,<3.0.0",
    "scikit-learn>=1.0.0,<2.0.0",
    "fastapi>=0.68.0,<1.0.0",
    "uvicorn[standard]>=0.15.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=5.4.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "httpx>=0.23.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
mlflow = [
    "mlflow>=2.0.0",
]

[project.scripts]
churn-train = "churn_mlops.models.train:main"
churn-serve = "churn_mlops.serving.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=churn_mlops --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
